Add information to mask gui
